Update to the new streamData endpoint
On Aug 27, Game Band consolidated streamGameData and other streams into a single endpoint, events/streamData. The old streamGameData data now lives in the value.games field. The lastUpdateTime field has been removed. BlaseballPlayback references lastUpdateTime, but I'm not totally sure what it's doing, and it seems to work ok without touching it, so I'm not going to mess with lastUpdateTime for now.